3. Przepisz program na Java https://dirask.com/posts/JavaScript-binary-search-algorithm-example-gp5lgD

+++ 13.03.2022-ZADANIE-3/src/Main.java
@@ -5,7 +5,7 @@ public class Main {
         int index = 0;
         int limit = array.length - 1;
         while (index <= limit){
-            double point = Math.ceil((index + limit) / 2);
+            double point = Math.ceil((float)(index + limit) / 2);
             int entry = array[(int) point];
             if (value > entry){
                 index = (int) (point + 1);
@@ -21,8 +21,7 @@ public class Main {
     }
 
     public static void main(String[] args) {
-
-	    int[] array = {4, 5, 7, 11, 12, 15, 15, 21, 40, 45};
+        int[] array = {4, 5, 7, 11, 12, 15, 15, 21, 40, 45};
         int index = searchIndex(array, 11);
         out.println(index);
     }
